A couple of things, read description
The ray class renamed a couple of innards to differ between primary and shadow rays. A couple of the debug views used the incorrect template bool being passed, additionally changed the ray-aabb function signature, to have an extra bool to indicate if the ray being tested is shadow ray or primary ray, to assist in counting.